profile config line load

+++ firmware/evo/Inc/shared.h
@@ -11,7 +11,7 @@
 
 void delay_ms(uint32_t amount);
 char* goto_next_arg(char* buf, char* buf_end);
-void strip_newline(char* line, uint8_t size);
+void strip_newline(char* line, uint32_t size);
 void idle_loop(void);
 
 #ifdef __cplusplus
